+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 ## Current Version: v2024-05-15.1.0
 
 ### Development Versions:
-- **[v2024-05-15.1.0]** Initial Commit
+- **[v2024-05-15.1.0]** MongoDb and OpenAPI Integration, Basic APIs